Implement GUI for the Terrain Generator. Also fix some cosmetic things.
Scaling: the model is now 10x10 Units large in both directions.
Translate: the camera and the Model are centered around the same point
Camera: because of the smaller scale, the camera can be nearer, and 
	the Zoom works much better.
GUI: JavaFX with FXML file and accompanying annotations.
Issues: the checkboxes for rivers, sites etc. are working and are passed
	into the mesh generating function, but are discarded there.